re-enabling tests ignored under #942
