+++ e2e/test/FaultInjectionPoolAmqpTests.MessageSendFaultInjectionPoolAmqpTests.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Microsoft.Azure.Devices.Client;
+using Microsoft.Azure.Devices.Client.Exceptions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
@@ -12,14 +13,14 @@ namespace Microsoft.Azure.Devices.E2ETests
 {
     public partial class FaultInjectionPoolAmqpTests : IDisposable
     {
-        private readonly string MessageReceive_DevicePrefix = $"E2E_MessageReceiveFaultInjectionPoolAmqpTests";
+        private readonly string MessageSend_DevicePrefix = $"E2E_MessageSendFaultInjectionPoolAmqpTests";
 
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_TcpConnectionLossReceiveRecovery_SingleConnection_Amqp()
+        public async Task Message_DeviceSak_TcpConnectionLossSendRecovery_SingleConnection_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -31,9 +32,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_TcpConnectionLossReceiveRecovery_SingleConnection_AmqpWs()
+        public async Task Message_DeviceSak_TcpConnectionLossSendRecovery_SingleConnection_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -45,9 +46,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_IoTHubSak_TcpConnectionLossReceiveRecovery_SingleConnection_Amqp()
+        public async Task Message_IoTHubSak_TcpConnectionLossSendRecovery_SingleConnection_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -60,9 +61,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_IoTHubSak_TcpConnectionLossReceiveRecovery_SingleConnection_AmqpWs()
+        public async Task Message_IoTHubSak_TcpConnectionLossSendRecovery_SingleConnection_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -75,9 +76,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpConnectionLossReceiveRecovery_SingleConnection_Amqp()
+        public async Task Message_DeviceSak_AmqpConnectionLossSendRecovery_SingleConnection_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -89,9 +90,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpConnectionLossReceiveRecovery_SingleConnection_AmqpWs()
+        public async Task Message_DeviceSak_AmqpConnectionLossSendRecovery_SingleConnection_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -103,9 +104,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_IotHubSak_AmqpConnectionLossReceiveRecovery_SingleConnection_Amqp()
+        public async Task Message_IotHubSak_AmqpConnectionLossSendRecovery_SingleConnection_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -118,9 +119,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_IoTHubSak_AmqpConnectionLossReceiveRecovery_SingleConnection_AmqpWs()
+        public async Task Message_IoTHubSak_AmqpConnectionLossSendRecovery_SingleConnection_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -131,13 +132,12 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
         // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpSessionLossReceiveRecovery_SingleConnection_Amqp()
+        public async Task Message_DeviceSak_AmqpSessionLossSendRecovery_SingleConnection_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -147,13 +147,12 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
         // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpSessionLossReceiveRecovery_SingleConnection_AmqpWs()
+        public async Task Message_DeviceSak_AmqpSessionLossSendRecovery_SingleConnection_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -163,13 +162,12 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
         // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [Ignore]
         [TestMethod]
-        public async Task Message_IoTHubSak_AmqpSessionLossReceiveRecovery_SingleConnection_Amqp()
+        public async Task Message_IoTHubSak_AmqpSessionLossSendRecovery_SingleConnection_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -180,13 +178,12 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
         // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [Ignore]
         [TestMethod]
-        public async Task Message_IoTHubSak_AmqpSessionLossReceiveRecovery_SingleConnection_AmqpWs()
+        public async Task Message_IoTHubSak_AmqpSessionLossSendRecovery_SingleConnection_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -197,63 +194,63 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
+        // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpC2DLinkDropReceiveRecovery_SingleConnection_Amqp()
+        public async Task Message_DeviceSak_AmqpD2CLinkDropSendRecovery_SingleConnection_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
                 PoolingOverAmqp.SingleConnection_DevicesCount,
-                FaultInjection.FaultType_AmqpC2D,
+                FaultInjection.FaultType_AmqpD2C,
                 FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
         }
 
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
+        // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpC2DLinkDropReceiveRecovery_SingleConnection_AmqpWs()
+        public async Task Message_DeviceSak_AmqpD2CLinkDropSendRecovery_SingleConnection_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
                 PoolingOverAmqp.SingleConnection_DevicesCount,
-                FaultInjection.FaultType_AmqpC2D,
+                FaultInjection.FaultType_AmqpD2C,
                 FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
         }
 
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
+        // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [Ignore]
         [TestMethod]
-        public async Task Message_IoTHubSak_AmqpC2DLinkDropReceiveRecovery_SingleConnection_Amqp()
+        public async Task Message_IoTHubSak_AmqpD2CLinkDropSendRecovery_SingleConnection_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
                 PoolingOverAmqp.SingleConnection_DevicesCount,
-                FaultInjection.FaultType_AmqpC2D,
+                FaultInjection.FaultType_AmqpD2C,
                 FaultInjection.FaultCloseReason_Boom,
                 authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
         }
 
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
+        // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [Ignore]
         [TestMethod]
-        public async Task Message_IoTHubSak_AmqpC2DLinkDropReceiveRecovery_SingleConnection_AmqpWs()
+        public async Task Message_IoTHubSak_AmqpD2CLinkDropSendRecovery_SingleConnection_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
                 PoolingOverAmqp.SingleConnection_DevicesCount,
-                FaultInjection.FaultType_AmqpC2D,
+                FaultInjection.FaultType_AmqpD2C,
                 FaultInjection.FaultCloseReason_Boom,
                 authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
         }
@@ -261,9 +258,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_GracefulShutdownReceiveRecovery_SingleConnection_Amqp()
+        public async Task Message_DeviceSak_GracefulShutdownSendRecovery_SingleConnection_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -275,9 +272,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_GracefulShutdownReceiveRecovery_SingleConnection_AmqpWs()
+        public async Task Message_DeviceSak_GracefulShutdownSendRecovery_SingleConnection_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -289,9 +286,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_IoTHubSak_GracefulShutdownReceiveRecovery_SingleConnection_Amqp()
+        public async Task Message_IoTHubSak_GracefulShutdownSendRecovery_SingleConnection_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -304,9 +301,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         // TODO: #943 - Honor different pool sizes for different connection pool settings.
         [Ignore]
         [TestMethod]
-        public async Task Message_IoTHubSak_GracefulShutdownReceiveRecovery_SingleConnection_AmqpWs()
+        public async Task Message_IoTHubSak_GracefulShutdownSendRecovery_SingleConnection_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.SingleConnection_PoolSize,
@@ -316,10 +313,130 @@ namespace Microsoft.Azure.Devices.E2ETests
                 authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
         }
 
+        // TODO: #943 - Honor different pool sizes for different connection pool settings.
+        [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_TcpConnectionLossReceiveRecovery_MultipleConnections_Amqp()
+        public async Task Message_DeviceSak_ThrottledConnectionRecovery_SingleConnection_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                PoolingOverAmqp.SingleConnection_PoolSize,
+                PoolingOverAmqp.SingleConnection_DevicesCount,
+                FaultInjection.FaultType_Throttle,
+                FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
+        }
+
+        // TODO: #943 - Honor different pool sizes for different connection pool settings.
+        [Ignore]
+        [TestMethod]
+        public async Task Message_DeviceSak_ThrottledConnectionRecovery_SingleConnection_AmqpWs()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                PoolingOverAmqp.SingleConnection_PoolSize,
+                PoolingOverAmqp.SingleConnection_DevicesCount,
+                FaultInjection.FaultType_Throttle,
+                FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
+        }
+
+        // TODO: #943 - Honor different pool sizes for different connection pool settings.
+        [Ignore]
+        [TestMethod]
+        public async Task Message_IoTHubSak_ThrottledConnectionRecovery_SingleConnection_Amqp()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                PoolingOverAmqp.SingleConnection_PoolSize,
+                PoolingOverAmqp.SingleConnection_DevicesCount,
+                FaultInjection.FaultType_Throttle,
+                FaultInjection.FaultCloseReason_Boom,
+                authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
+        }
+
+        // TODO: #943 - Honor different pool sizes for different connection pool settings.
+        [Ignore]
+        [TestMethod]
+        public async Task Message_IoTHubSak_ThrottledConnectionRecovery_SingleConnection_AmqpWs()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                PoolingOverAmqp.SingleConnection_PoolSize,
+                PoolingOverAmqp.SingleConnection_DevicesCount,
+                FaultInjection.FaultType_Throttle,
+                FaultInjection.FaultCloseReason_Boom,
+                authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
+        }
+
+        // TODO: #943 - Honor different pool sizes for different connection pool settings.
+        [Ignore]
+        [TestMethod]
+        [ExpectedException(typeof(UnauthorizedException))]
+        public async Task Message_DeviceSak_AuthenticationNoRecovery_SingleConnection_Amqp()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                PoolingOverAmqp.SingleConnection_PoolSize,
+                PoolingOverAmqp.SingleConnection_DevicesCount,
+                FaultInjection.FaultType_Auth,
+                FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
+        }
+
+        // TODO: #943 - Honor different pool sizes for different connection pool settings.
+        [Ignore]
+        [TestMethod]
+        [ExpectedException(typeof(UnauthorizedException))]
+        public async Task Message_DeviceSak_AuthenticationNoRecovery_SingleConnection_AmqpWs()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                PoolingOverAmqp.SingleConnection_PoolSize,
+                PoolingOverAmqp.SingleConnection_DevicesCount,
+                FaultInjection.FaultType_Auth,
+                FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
+        }
+
+        // TODO: #943 - Honor different pool sizes for different connection pool settings.
+        [Ignore]
+        [TestMethod]
+        [ExpectedException(typeof(UnauthorizedException))]
+        public async Task Message_IoTHubSak_AuthenticationNoRecovery_SingleConnection_Amqp()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                PoolingOverAmqp.SingleConnection_PoolSize,
+                PoolingOverAmqp.SingleConnection_DevicesCount,
+                FaultInjection.FaultType_Auth,
+                FaultInjection.FaultCloseReason_Boom,
+                authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
+        }
+
+        // TODO: #943 - Honor different pool sizes for different connection pool settings.
+        [Ignore]
+        [TestMethod]
+        [ExpectedException(typeof(UnauthorizedException))]
+        public async Task Message_IoTHubSak_AuthenticationNoRecovery_SingleConnection_AmqpWs()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                PoolingOverAmqp.SingleConnection_PoolSize,
+                PoolingOverAmqp.SingleConnection_DevicesCount,
+                FaultInjection.FaultType_Auth,
+                FaultInjection.FaultCloseReason_Boom,
+                authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_DeviceSak_TcpConnectionLossSendRecovery_MultipleConnections_Amqp()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -329,9 +446,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        public async Task Message_DeviceSak_TcpConnectionLossReceiveRecovery_MultipleConnections_AmqpWs()
+        public async Task Message_DeviceSak_TcpConnectionLossSendRecovery_MultipleConnections_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -341,9 +458,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        public async Task Message_IoTHubSak_TcpConnectionLossReceiveRecovery_MultipleConnections_Amqp()
+        public async Task Message_IoTHubSak_TcpConnectionLossSendRecovery_MultipleConnections_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -354,9 +471,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        public async Task Message_IoTHubSak_TcpConnectionLossReceiveRecovery_MultipleConnections_AmqpWs()
+        public async Task Message_IoTHubSak_TcpConnectionLossSendRecovery_MultipleConnections_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -367,9 +484,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpConnectionLossReceiveRecovery_MultipleConnections_Amqp()
+        public async Task Message_DeviceSak_AmqpConnectionLossSendRecovery_MultipleConnections_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -379,9 +496,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpConnectionLossReceiveRecovery_MultipleConnections_AmqpWs()
+        public async Task Message_DeviceSak_AmqpConnectionLossSendRecovery_MultipleConnections_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -391,9 +508,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        public async Task Message_IotHubSak_AmqpConnectionLossReceiveRecovery_MultipleConnections_Amqp()
+        public async Task Message_IotHubSak_AmqpConnectionLossSendRecovery_MultipleConnections_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -404,9 +521,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        public async Task Message_IoTHubSak_AmqpConnectionLossReceiveRecovery_MultipleConnections_AmqpWs()
+        public async Task Message_IoTHubSak_AmqpConnectionLossSendRecovery_MultipleConnections_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -416,13 +533,11 @@ namespace Microsoft.Azure.Devices.E2ETests
                 authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
         }
 
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
         // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
-        [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpSessionLossReceiveRecovery_MultipleConnections_Amqp()
+        public async Task Message_DeviceSak_AmqpSessionLossSendRecovery_MultipleConnections_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -431,13 +546,11 @@ namespace Microsoft.Azure.Devices.E2ETests
                 "").ConfigureAwait(false);
         }
 
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
         // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
-        [Ignore]
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpSessionLossReceiveRecovery_MultipleConnections_AmqpWs()
+        public async Task Message_DeviceSak_AmqpSessionLossSendRecovery_MultipleConnections_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -446,13 +559,11 @@ namespace Microsoft.Azure.Devices.E2ETests
                 "").ConfigureAwait(false);
         }
 
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
         // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
-        [Ignore]
         [TestMethod]
-        public async Task Message_IoTHubSak_AmqpSessionLossReceiveRecovery_MultipleConnections_Amqp()
+        public async Task Message_IoTHubSak_AmqpSessionLossSendRecovery_MultipleConnections_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -462,13 +573,11 @@ namespace Microsoft.Azure.Devices.E2ETests
                 authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
         }
 
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
         // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
-        [Ignore]
         [TestMethod]
-        public async Task Message_IoTHubSak_AmqpSessionLossReceiveRecovery_MultipleConnections_AmqpWs()
+        public async Task Message_IoTHubSak_AmqpSessionLossSendRecovery_MultipleConnections_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -478,68 +587,64 @@ namespace Microsoft.Azure.Devices.E2ETests
                 authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
         }
 
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
-        [Ignore]
+        // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpC2DLinkDropReceiveRecovery_MultipleConnections_Amqp()
+        public async Task Message_DeviceSak_AmqpD2CLinkDropSendRecovery_MultipleConnections_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
                 PoolingOverAmqp.MultipleConnections_DevicesCount,
-                FaultInjection.FaultType_AmqpC2D,
+                FaultInjection.FaultType_AmqpD2C,
                 FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
         }
 
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
-        [Ignore]
+        // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [TestMethod]
-        public async Task Message_DeviceSak_AmqpC2DLinkDropReceiveRecovery_MultipleConnections_AmqpWs()
+        public async Task Message_DeviceSak_AmqpD2CLinkDropSendRecovery_MultipleConnections_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
                 PoolingOverAmqp.MultipleConnections_DevicesCount,
-                FaultInjection.FaultType_AmqpC2D,
+                FaultInjection.FaultType_AmqpD2C,
                 FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
         }
 
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
-        [Ignore]
+        // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [TestMethod]
-        public async Task Message_IoTHubSak_AmqpC2DLinkDropReceiveRecovery_MultipleConnections_Amqp()
+        public async Task Message_IoTHubSak_AmqpD2CLinkDropSendRecovery_MultipleConnections_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
                 PoolingOverAmqp.MultipleConnections_DevicesCount,
-                FaultInjection.FaultType_AmqpC2D,
+                FaultInjection.FaultType_AmqpD2C,
                 FaultInjection.FaultCloseReason_Boom,
                 authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
         }
 
-        // TODO: #942 - NullReferenceException is thrown while disposing message, since fault injection causes ReceivingAmqpLink to be null.
-        [Ignore]
+        // TODO: #950 - Link/session faults for message send/ method/ twin operations closes the connection.
         [TestMethod]
-        public async Task Message_IoTHubSak_AmqpC2DLinkDropReceiveRecovery_MultipleConnections_AmqpWs()
+        public async Task Message_IoTHubSak_AmqpD2CLinkDropSendRecovery_MultipleConnections_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
                 PoolingOverAmqp.MultipleConnections_DevicesCount,
-                FaultInjection.FaultType_AmqpC2D,
+                FaultInjection.FaultType_AmqpD2C,
                 FaultInjection.FaultCloseReason_Boom,
                 authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
         }
 
         [TestMethod]
-        public async Task Message_DeviceSak_GracefulShutdownReceiveRecovery_MultipleConnections_Amqp()
+        public async Task Message_DeviceSak_GracefulShutdownSendRecovery_MultipleConnections_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -549,9 +654,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        public async Task Message_DeviceSak_GracefulShutdownReceiveRecovery_MultipleConnections_AmqpWs()
+        public async Task Message_DeviceSak_GracefulShutdownSendRecovery_MultipleConnections_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -561,9 +666,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        public async Task Message_IoTHubSak_GracefulShutdownReceiveRecovery_MultipleConnections_Amqp()
+        public async Task Message_IoTHubSak_GracefulShutdownSendRecovery_MultipleConnections_Amqp()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_Tcp_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -574,9 +679,9 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        public async Task Message_IoTHubSak_GracefulShutdownReceiveRecovery_MultipleConnections_AmqpWs()
+        public async Task Message_IoTHubSak_GracefulShutdownSendRecovery_MultipleConnections_AmqpWs()
         {
-            await ReceiveMessageRecoveryPoolOverAmqp(
+            await SendMessageRecoveryPoolOverAmqp(
                 TestDeviceType.Sasl,
                 Client.TransportType.Amqp_WebSocket_Only,
                 PoolingOverAmqp.MultipleConnections_PoolSize,
@@ -586,7 +691,111 @@ namespace Microsoft.Azure.Devices.E2ETests
                 authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
         }
 
-        private async Task ReceiveMessageRecoveryPoolOverAmqp(
+        [TestMethod]
+        public async Task Message_DeviceSak_ThrottledConnectionRecovery_MultipleConnections_Amqp()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                PoolingOverAmqp.MultipleConnections_PoolSize,
+                PoolingOverAmqp.MultipleConnections_DevicesCount,
+                FaultInjection.FaultType_Throttle,
+                FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_DeviceSak_ThrottledConnectionRecovery_MultipleConnections_AmqpWs()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                PoolingOverAmqp.MultipleConnections_PoolSize,
+                PoolingOverAmqp.MultipleConnections_DevicesCount,
+                FaultInjection.FaultType_Throttle,
+                FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_IoTHubSak_ThrottledConnectionRecovery_MultipleConnections_Amqp()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                PoolingOverAmqp.MultipleConnections_PoolSize,
+                PoolingOverAmqp.MultipleConnections_DevicesCount,
+                FaultInjection.FaultType_Throttle,
+                FaultInjection.FaultCloseReason_Boom,
+                authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_IoTHubSak_ThrottledConnectionRecovery_MultipleConnections_AmqpWs()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                PoolingOverAmqp.MultipleConnections_PoolSize,
+                PoolingOverAmqp.MultipleConnections_DevicesCount,
+                FaultInjection.FaultType_Throttle,
+                FaultInjection.FaultCloseReason_Boom,
+                authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        [ExpectedException(typeof(UnauthorizedException))]
+        public async Task Message_DeviceSak_AuthenticationNoRecovery_MultipleConnections_Amqp()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                PoolingOverAmqp.MultipleConnections_PoolSize,
+                PoolingOverAmqp.MultipleConnections_DevicesCount,
+                FaultInjection.FaultType_Auth,
+                FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        [ExpectedException(typeof(UnauthorizedException))]
+        public async Task Message_DeviceSak_AuthenticationNoRecovery_MultipleConnections_AmqpWs()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                PoolingOverAmqp.MultipleConnections_PoolSize,
+                PoolingOverAmqp.MultipleConnections_DevicesCount,
+                FaultInjection.FaultType_Auth,
+                FaultInjection.FaultCloseReason_Boom).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        [ExpectedException(typeof(UnauthorizedException))]
+        public async Task Message_IoTHubSak_AuthenticationNoRecovery_MultipleConnections_Amqp()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                PoolingOverAmqp.MultipleConnections_PoolSize,
+                PoolingOverAmqp.MultipleConnections_DevicesCount,
+                FaultInjection.FaultType_Auth,
+                FaultInjection.FaultCloseReason_Boom,
+                authScope:ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        [ExpectedException(typeof(UnauthorizedException))]
+        public async Task Message_IoTHubSak_AuthenticationNoRecovery_MultipleConnections_AmqpWs()
+        {
+            await SendMessageRecoveryPoolOverAmqp(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                PoolingOverAmqp.MultipleConnections_PoolSize,
+                PoolingOverAmqp.MultipleConnections_DevicesCount,
+                FaultInjection.FaultType_Auth,
+                FaultInjection.FaultCloseReason_Boom,
+                authScope: ConnectionStringAuthScope.IoTHub).ConfigureAwait(false);
+        }
+
+        private async Task SendMessageRecoveryPoolOverAmqp(
             TestDeviceType type,
             Client.TransportType transport,
             int poolSize,
@@ -597,25 +806,35 @@ namespace Microsoft.Azure.Devices.E2ETests
             int durationInSec = FaultInjection.DefaultDurationInSec,
             ConnectionStringAuthScope authScope = ConnectionStringAuthScope.Device)
         {
-            // Initialize the service client
-            ServiceClient serviceClient = ServiceClient.CreateFromConnectionString(Configuration.IoTHub.ConnectionString);
+            Dictionary<string, EventHubTestListener> eventHubListeners = new Dictionary<string, EventHubTestListener>();
+
+            Func<DeviceClient, TestDevice, Task> initOperation = async (deviceClient, testDevice) =>
+            {
+                EventHubTestListener testListener = await EventHubTestListener.CreateListener(testDevice.Id).ConfigureAwait(false);
+                eventHubListeners.Add(testDevice.Id, testListener);
+            };
 
             Func<DeviceClient, TestDevice, Task> testOperation = async (deviceClient, testDevice) =>
             {
-                (Message msg, string messageId, string payload, string p1Value) = MessageReceiveE2ETests.ComposeC2DTestMessage();
-                _log.WriteLine($"{nameof(FaultInjectionPoolAmqpTests)}: Sending message to device {testDevice.Id}: payload='{payload}' p1Value='{p1Value}'");
-                await serviceClient.SendAsync(testDevice.Id, msg).ConfigureAwait(false);
-
-                _log.WriteLine($"{nameof(FaultInjectionPoolAmqpTests)}: Preparing to receive message for device {testDevice.Id}");
+                _log.WriteLine($"{nameof(FaultInjectionPoolAmqpTests)}: Preparing to send message for device {testDevice.Id}");
                 await deviceClient.OpenAsync().ConfigureAwait(false);
-                await MessageReceiveE2ETests.VerifyReceivedC2DMessageAsync(transport, deviceClient, testDevice.Id, payload, p1Value).ConfigureAwait(false);
+
+                (Client.Message testMessage, string messageId, string payload, string p1Value) = MessageSendE2ETests.ComposeD2CTestMessage();
+
+                _log.WriteLine($"{nameof(FaultInjectionPoolAmqpTests)}.{testDevice.Id}: payload='{payload}' p1Value='{p1Value}'");
+                await deviceClient.SendEventAsync(testMessage).ConfigureAwait(false);
+
+                EventHubTestListener testListener = eventHubListeners[testDevice.Id];
+                bool isReceived = await testListener.WaitForMessage(testDevice.Id, payload, p1Value).ConfigureAwait(false);
+                Assert.IsTrue(isReceived, $"Message is not received for device {testDevice.Id}.");
             };
 
             Func<IList<DeviceClient>, Task> cleanupOperation = async (deviceClients) =>
             {
-                await serviceClient.CloseAsync().ConfigureAwait(false);
-                serviceClient.Dispose();
-
+                foreach (var listener in eventHubListeners)
+                {
+                    await listener.Value.CloseAsync().ConfigureAwait(false);
+                }
                 foreach (DeviceClient deviceClient in deviceClients)
                 {
                     deviceClient.Dispose();
@@ -623,7 +842,7 @@ namespace Microsoft.Azure.Devices.E2ETests
             };
 
             await FaultInjectionPoolingOverAmqp.TestFaultInjectionPoolAmqpAsync(
-                MessageReceive_DevicePrefix,
+                MessageSend_DevicePrefix,
                 transport,
                 poolSize,
                 devicesCount,
@@ -631,7 +850,7 @@ namespace Microsoft.Azure.Devices.E2ETests
                 reason,
                 delayInSec,
                 durationInSec,
-                (d, t) => { return Task.FromResult(false);  },
+                initOperation,
                 testOperation,
                 cleanupOperation,
                 authScope).ConfigureAwait(false);
