+++ e2e/test/FaultInjectionPoolAmqpTests.MethodFaultInjectionPoolAmqpTests.cs
