+++ vsts/test-release-nuget.yaml
@@ -77,12 +77,7 @@ phases:
     displayName: build
     env:
       IOTHUB_CONN_STRING_CSHARP: $(IOTHUB-CONN-STRING-CSHARP)
-      IOTHUB_DEVICE_CONN_STRING: $(IOTHUB-DEVICE-CONN-STRING)
-      IOTHUB_DEVICE_CONN_STRING2: $(IOTHUB-DEVICE-CONN-STRING2)
-      IOTHUB_MODULE_CONN_STRING: $(IOTHUB-MODULE-CONN-STRING)
-      IOTHUB_MODULE_CONN_STRING2: $(IOTHUB-MODULE-CONN-STRING2)
       IOTHUB_PFX_X509_THUMBPRINT: $(IOTHUB-PFX-X509-THUMBPRINT)
-      IOTHUB_PFX_X509_THUMBPRINT2: $(IOTHUB-PFX-X509-THUMBPRINT2)
       IOTHUB_X509_PFX_CERTIFICATE: $(IOTHUB-X509-PFX-CERTIFICATE)
       IOTHUB_EVENTHUB_CONN_STRING_CSHARP: $(IOTHUB-EVENTHUB-CONN-STRING-CSHARP)
       IOTHUB_EVENTHUB_COMPATIBLE_NAME: $(IOTHUB-EVENTHUB-COMPATIBLE-NAME)
@@ -193,12 +188,7 @@ phases:
     displayName: build
     env:
       IOTHUB_CONN_STRING_CSHARP: $(IOTHUB-CONN-STRING-CSHARP)
-      IOTHUB_DEVICE_CONN_STRING: $(IOTHUB-DEVICE-CONN-STRING)
-      IOTHUB_DEVICE_CONN_STRING2: $(IOTHUB-DEVICE-CONN-STRING2)
-      IOTHUB_MODULE_CONN_STRING: $(IOTHUB-MODULE-CONN-STRING)
-      IOTHUB_MODULE_CONN_STRING2: $(IOTHUB-MODULE-CONN-STRING2)
       IOTHUB_PFX_X509_THUMBPRINT: $(IOTHUB-PFX-X509-THUMBPRINT)
-      IOTHUB_PFX_X509_THUMBPRINT2: $(IOTHUB-PFX-X509-THUMBPRINT2)
       IOTHUB_X509_PFX_CERTIFICATE: $(IOTHUB-X509-PFX-CERTIFICATE)
       IOTHUB_EVENTHUB_CONN_STRING_CSHARP: $(IOTHUB-EVENTHUB-CONN-STRING-CSHARP)
       IOTHUB_EVENTHUB_COMPATIBLE_NAME: $(IOTHUB-EVENTHUB-COMPATIBLE-NAME)
